Decode the README as utf-8
tox no longer passed environment variables from the calling environment.
This causes problems, because the LANG environment variable is how
Python attempts to figure out the locale, and defaults to ASCII. The
README file has utf-8 characters.

This means that right now this package cannot be installed under tox 2
with Python 3, unless tox in configured to pass that environment
variable from the context.