Use tox for testing on Travis
Put all the testing configuration and various python versions to test in
tox.ini, and use TOXENV to choose which one to build on Travis. Use
Travis' build matrix to separate the runs for all of the tox envs.

The sudo: false line should speed up the builds, because it allows
Travis to use containers.